zaqaz/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13.0,<4.0
djangorestframework-simplejwt>=5.0.0
django-debug-toolbar
django-redis>=5.0.0
